feat: enhance Docker setup with PM2 integration and add backup functionality
