+++ .dockerignore
@@ -21,4 +21,3 @@ dist
 *.swo
 uptime-kuma
 scripts
-backup.js
