+++ ecosystem.config.js
@@ -1,0 +1,24 @@
+module.exports = {
+  apps: [
+    {
+      name: 'erazor-server',
+      script: './dist/main.js',
+      instances: 'max', // Use all available CPU cores
+      exec_mode: 'cluster',
+      env: {
+        NODE_ENV: 'production',
+        PORT: process.env.PORT || 4545
+      },
+      max_memory_restart: '1G',
+      node_args: '--max-old-space-size=1024',
+      autorestart: true,
+      watch: false,
+      max_restarts: 10,
+      min_uptime: '10s',
+      error_file: './logs/err.log',
+      out_file: './logs/out.log',
+      log_file: './logs/combined.log',
+      time: true
+    }
+  ]
+};
